Fix bug related to native method string return type
Added a property to `LibImportAttribute` to specify a return type for a native method that returns a pointer `nint`.
In the case of a `nint` that is intended to be a string, the native method is called in `Marshal.PtrToStringAuto`, and the return type of the generated managed method is set to the type set in the `ReturnType` field of `LibImportAttribute`.

Before this change, calling a native method which returns a string would result in a call stack like this:
```
ntdll.dll!RtlReportCriticalFailure()
ntdll.dll!RtlpHeapHandleError()
ntdll.dll!RtlpHpHeapHandleError()
ntdll.dll!RtlpLogHeapFailure()
ntdll.dll!RtlpFreeHeapInternal()
ntdll.dll!RtlFreeHeap()
```